move module favorite screen
